+++ src/main/java/com/ph/phbackend/payload/request/PatientRequest.java
@@ -4,15 +4,15 @@ public class PatientRequest {
 
     private String firstName;
     private String lastName;
-    private Integer weight;
-    private Integer height;
+    private Long weight;
+    private Long height;
     private String age;
     private String gender;
     private Long userId;
 
     public PatientRequest() {}
 
-    public PatientRequest(String firstName, String lastName, Integer weight, Integer height, String age, String gender, Long userId) {
+    public PatientRequest(String firstName, String lastName, Long weight, Long height, String age, String gender, Long userId) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.weight = weight;
@@ -38,19 +38,19 @@ public class PatientRequest {
         this.lastName = lastName;
     }
 
-    public Integer getWeight() {
+    public Long getWeight() {
         return weight;
     }
 
-    public void setWeight(Integer weight) {
+    public void setWeight(Long weight) {
         this.weight = weight;
     }
 
-    public Integer getHeight() {
+    public Long getHeight() {
         return height;
     }
 
-    public void setHeight(Integer height) {
+    public void setHeight(Long height) {
         this.height = height;
     }
